+++ src/strongbox/config.clj
@@ -110,6 +110,15 @@
     (assoc cfg :catalogue-location-list -default-catalogue-list--v2)
     cfg))
 
+(defn-spec handle-column-preferences map?
+  "handles upgrading of the default column list.
+  if the config is using the v1 defaults, upgrade to v2 defaults."
+  [cfg map?]
+  (if (= (-> cfg :preferences :ui-selected-columns set)
+         (set sp/default-column-list--v1))
+    (assoc-in cfg [:preferences :ui-selected-columns] sp/default-column-list--v2)
+    cfg))
+
 (defn remove-invalid-addon-dirs
   "removes any `addon-dir-map` items from the given configuration whose directories do not exist"
   [cfg]
@@ -150,6 +159,7 @@
                 handle-selected-addon-dir
                 remove-invalid-catalogue-location-entries
                 add-github-catalogue
+                handle-column-preferences
                 strip-unspecced-keys)
         message (format "configuration from %s is invalid and will be ignored: %s"
                         msg (s/explain-str ::sp/user-config cfg))]